feat(storage-provider): Move benchmark pallet and config to a separate module
