+++ pallets/storage-provider/benchmarks/src/lib.rs
@@ -5,11 +5,14 @@ extern crate alloc;
 #[cfg(feature = "runtime-benchmarks")]
 pub mod inner;
 
-#[cfg(feature = "runtime-benchmarks")]
-pub use inner::*;
-
 #[cfg(all(feature = "runtime-benchmarks", test))]
 pub(crate) mod mock;
 
 #[cfg(any(feature = "runtime-benchmarks", test))]
 mod utils;
+
+#[cfg(any(feature = "runtime-benchmarks", test))]
+mod pallet;
+
+#[cfg(feature = "runtime-benchmarks")]
+pub use pallet::*;
